+++ test/runner_test.rb
@@ -55,7 +55,9 @@ module Byebug
 
     def test_run_with_an_nonexistent_script
       with_command_line('bin/byebug', 'non_existent_script.rb') do
-        assert_raises(Runner::NonExistentScript) { runner.run }
+        runner.run
+
+        check_error_includes "The script doesn't exist"
       end
     end